seperated valuesScreens into files
